telemonitoramento-web/requirements.txt
streamlit
psycopg2-binary
pandas
matplotlib
python-dotenv
plotly
cryptography
reportlab
streamlit-extras
pytest
pytest-cov
